리팩토링

+++ src/main/java/kr/co/torpedo/webservicemanager/controller/UserController.java
@@ -3,6 +3,8 @@ package kr.co.torpedo.webservicemanager.controller;
 import java.util.Map;
 
 import javax.inject.Inject;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpSession;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -24,31 +26,55 @@ public class UserController {
 	private UserService userService;
 
 	@RequestMapping(value = "/register", method = RequestMethod.GET)
-	public String register() {
+	public String register(HttpServletRequest httpServletRequest) {
 		logger.info("register");
+		if (!checkSession(httpServletRequest)) {
+			logger.info("Don't have session");
+			return "redirect:/";
+		}
 		return "register";
 	}
 
 	@RequestMapping(value = "/detail", method = RequestMethod.GET)
-	public String detail(@RequestParam("id") int id, Model model) {
+	public String detail(@RequestParam("id") int id, Model model, HttpServletRequest httpServletRequest) {
 		logger.info("detail");
+		if (!checkSession(httpServletRequest)) {
+			logger.info("Don't have session");
+			return "redirect:/";
+		}
 		model.addAttribute("User", userService.selectUser(id));
 		return "detail";
 	}
 
 	@RequestMapping(value = "/update", method = RequestMethod.GET)
-	public String update(@RequestParam("id") int id, Model model) {
+	public String update(@RequestParam("id") int id, Model model, HttpServletRequest httpServletRequest) {
 		logger.info("update");
+		if (!checkSession(httpServletRequest)) {
+			logger.info("Don't have session");
+			return "redirect:/";
+		}
 		model.addAttribute("User", userService.selectUser(id));
 		return "update";
 	}
 
 	@ResponseBody
 	@RequestMapping(value = "/delete", method = RequestMethod.POST)
-	public String delete(@RequestBody Map<String, Object> params) {
+	public String delete(@RequestBody Map<String, Object> params, HttpServletRequest httpServletRequest) {
 		logger.info("delete");
+		if (!checkSession(httpServletRequest)) {
+			logger.info("Don't have session");
+			return "redirect:/";
+		}
 		userService.delete(Integer.parseInt(params.get("id").toString()));
 		logger.info("delete success");
 		return "redirect:/list";
+	}
+
+	public boolean checkSession(HttpServletRequest httpServletRequest) {
+		HttpSession session = httpServletRequest.getSession();
+		if (session.getAttribute("Admin") == null) {
+			return false;
+		}
+		return true;
 	}
 }
